refactor(cmd): move manager command into its own package
cmd/main.go is now a thin entrypoint: it keeps only the ldflag-injected version
vars (so -X main.version=... stays valid) and assembles the cobra root with the
manager and extension-server subcommands.

The controller-manager command moves to internal/cmd/manager (package
managercmd) as NewCommand(BuildInfo), mirroring how the extension server lives
in internal/extensionserver/cmd. The scheme/init and the
initializeClusterDiscovery/ignoreCanceled helpers move with it. Build info is
passed in via a BuildInfo struct instead of package-level vars.
